+++ Platform/Vulkan/DisplaySurface.cpp
@@ -187,7 +187,7 @@ void DisplaySurface::GetQueues()
 	if (!deviceManager)
 		return;
 	const std::vector<vk::QueueFamilyProperties> &queue_props=deviceManager->GetQueueProperties();
-	uint32_t queue_family_count=queue_props.size();
+	uint32_t queue_family_count=(uint32_t)queue_props.size();
 	// Iterate over each queue to learn whether it supports presenting:
 	std::vector<vk::Bool32> supportsPresent(queue_family_count);
 	for (uint32_t i = 0; i < queue_family_count; i++)
